+++ app/scripts/HorizontalTiled1DPixiTrack.js
@@ -9,7 +9,7 @@ import { colorToHex, showMousePosition } from './utils';
 class HorizontalTiled1DPixiTrack extends Tiled1DPixiTrack {
   constructor(context, options) {
     super(context, options);
-    const { animate } = context;
+    const { animate, isShowGlobalMousePosition } = context;
 
     this.constIndicator = new PIXI.Graphics();
     this.pMain.addChild(this.constIndicator);
@@ -19,11 +19,14 @@ class HorizontalTiled1DPixiTrack extends Tiled1DPixiTrack {
 
     this.animate = animate;
     this.options = options;
+    this.isShowGlobalMousePosition = isShowGlobalMousePosition;
 
     this.pubSubs = [];
 
     if (this.options.showMousePosition && !this.hideMousePosition) {
-      this.hideMousePosition = showMousePosition(this, this.is2d);
+      this.hideMousePosition = showMousePosition(
+        this, this.is2d, this.isShowGlobalMousePosition()
+      );
     }
   }
 
@@ -38,7 +41,9 @@ class HorizontalTiled1DPixiTrack extends Tiled1DPixiTrack {
     this.options = options;
 
     if (this.options.showMousePosition && !this.hideMousePosition) {
-      this.hideMousePosition = showMousePosition(this);
+      this.hideMousePosition = showMousePosition(
+        this, this.is2d, this.isShowGlobalMousePosition()
+      );
     }
 
     if (!this.options.showMousePosition && this.hideMousePosition) {